requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Limiter==3.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai
flask-cors
